Migration of Pregame into src and include foldeer
